video_verifier/backend/requirement.txt
Flask
Flask-Cors
python-dotenv
requests
opencv-python # Replaces pyexiftool
hachoir       # For metadata extraction (replaces ffprobe/exiftool)
# Google Cloud dependencies
google-cloud-speech
google-api-python-client
google-auth
google-generativeai>=0.3.0  # For Gemini API integration
# For audio extraction
moviepy
# For image hosting
cloudinary
# For NLP tasks
spacy>=3.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
# For visualization
matplotlib
scipy
pytesseract
pytube
yt-dlp
azure-ai-vision-imageanalysis